pyproject.toml
[project]
name = "qwen-image-edit-server"
version = "0.1.0"
description = "OpenAI-compatible API for Qwen-Image-Edit-2509"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "torch>=2.1.2",
    "diffusers>=0.30.0",
    "transformers>=4.40.0",
    "accelerate>=0.30.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.9",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
